# src/main/generated/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        owner:
          type: string
        shares:
          type: integer
          format: int32
        symbol:
          type: string
        notional:
          type: number
          format: float
        price:
          type: number
          format: float
        commission:
          type: number
          format: float
        _id:
          type: string
        when:
          type: string
          format: date-time
          examples:
          - 2022-03-10T16:15:50Z
    Transactions:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
paths:
...
  /trade-history/trades/{owner}:
    get:
      summary: Get trade history of specified owner
      description: Get an array of owner's transactions
      parameters:
      - description: Owner name
        required: true
        name: owner
        in: path
        schema:
          type: string
      responses:
        "404":
          description: The Mongo database cannot be found.
          content:
            text/plain: {}
        "200":
          description: The trades for the requested owner have been retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transactions"
  /trade-history/trades/{owner}/{symbol}:
    get:
      summary: Get trade history of specified owner for the specified stock symbol
      description: Get an array of the owner's transactions for the specified stock
        symbol
      parameters:
      - description: Owner name
        required: true
        name: owner
        in: path
        schema:
          type: string
      - description: Symbol name
        required: true
        name: symbol
        in: path
        schema:
          type: string
      responses:
        "404":
          description: The Mongo database cannot be found.
          content:
            text/plain: {}
        "200":
          description: The trades for the requested owner and symbol have been retrieved
            successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transactions"
info:
  title: Generated API
  version: "1.0"