# src/main/java/com/ibm/hybrid/cloud/sample/stocktrader/tradehistory/rest/Trades.java
/*
       Copyright 2018-2021 IBM Corp All Rights Reserved
       Copyright 2022-2025 Kyndryl, All Rights Reserved
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
   You may obtain a copy of the License at
       http://www.apache.org/licenses/LICENSE-2.0
   Unless required by applicable law or agreed to in writing, software
   distributed under the License is distributed on an "AS IS" BASIS,
   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   See the License for the specific language governing permissions and
   limitations under the License.
 */

package com.ibm.hybrid.cloud.sample.stocktrader.tradehistory.rest;

import com.ibm.hybrid.cloud.sample.stocktrader.tradehistory.mongo.MongoConnector;
import io.opentelemetry.instrumentation.annotations.WithSpan;
import io.swagger.annotations.Api;
import jakarta.enterprise.context.RequestScoped;
import jakarta.inject.Inject;
import jakarta.ws.rs.*;
import jakarta.ws.rs.core.MediaType;
import org.eclipse.microprofile.openapi.annotations.OpenAPIDefinition;
import org.eclipse.microprofile.openapi.annotations.Operation;
import org.eclipse.microprofile.openapi.annotations.info.Contact;
import org.eclipse.microprofile.openapi.annotations.info.Info;
import org.eclipse.microprofile.openapi.annotations.info.License;
import org.eclipse.microprofile.openapi.annotations.media.Content;
import org.eclipse.microprofile.openapi.annotations.media.Schema;
import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;

@Path("/")
@Api(tags = {"trade-history"})
@Produces("application/json")
@OpenAPIDefinition(
    info = @Info(
        title = "Trade History",
        version = "1.0",
        description = "TradeHistory API",
        contact = @Contact(url = "https://github.com/IBMStockTrader", name = "IBMStockTrader"),
        license = @License(name = "License", url = "https://github.com/IBMStockTrader/trade-history/blob/master/LICENSE")
    )
)
@RequestScoped
public class Trades {

    @Inject
    private MongoConnector mConnector;

    @Path("/latestBuy")
    @GET
    @Produces(MediaType.APPLICATION_JSON)
    @APIResponses(value = {
        @APIResponse(
            responseCode = "404",
            description = "The Mongo database cannot be found. ",
            content = @Content(
                mediaType = "text/plain")),
        @APIResponse(
            responseCode = "200",
            description = "The latest trade has been retrieved successfully.",
            content = @Content(
                mediaType = "application/json",
                schema = @Schema(implementation = Transaction.class)))})
    @Operation(
        summary = "Shows the latest trade.",
        description = "Retrieve the latest record from the mongo database."
    )
    @WithSpan
    public Transaction latestBuy() {
        Transaction trade = null;

        if (mConnector != null) {
            trade = mConnector.getLatestBuy();
        }

        return trade;
    }

    @Path("/trades/{owner}")
    @GET
    @Produces(MediaType.APPLICATION_JSON)
    @APIResponses(value = {
        @APIResponse(
            responseCode = "404",
            description = "The Mongo database cannot be found.",
            content = @Content(
                mediaType = "text/plain")),
        @APIResponse(
            responseCode = "200",
            description = "The trades for the requested owner have been retrieved successfully.",
            content = @Content(
                mediaType = "application/json",
                schema = @Schema(implementation = Transactions.class)))})
    @Operation(summary = "Get trade history of specified owner",
        description = "Get an array of owner's transactions")
    @WithSpan
    public Transactions getTradesByOwner(@Parameter(description = "Owner name", required = true) @PathParam("owner") String ownerName) {

        Transactions trades = null;

        if (mConnector != null) {
            trades = mConnector.getTrades(ownerName);
        }

        return trades;
    }

    @Path("/trades/{owner}/{symbol}")
    @GET
    @Produces(MediaType.APPLICATION_JSON)
    @APIResponses(value = {
        @APIResponse(
            responseCode = "404",
            description = "The Mongo database cannot be found.",
            content = @Content(
                mediaType = "text/plain")),
        @APIResponse(
            responseCode = "200",
            description = "The trades for the requested owner and symbol have been retrieved successfully.",
            content = @Content(
                mediaType = "application/json",
                schema = @Schema(implementation = Transactions.class)))})
    @Operation(summary = "Get trade history of specified owner for the specified stock symbol",
        description = "Get an array of the owner's transactions for the specified stock symbol")
    @WithSpan
    public Transactions getTradesByOwnerAndSymbol(
        @Parameter(description = "Owner name", required = true) @PathParam("owner") String ownerName,
        @Parameter(description = "Symbol name", required = true) @PathParam("symbol") String symbol) {

        Transactions trades = null;

        if (mConnector != null) {
            trades = mConnector.getTradesForSymbol(ownerName, symbol);
        }

        return trades;
    }

    @Path("/shares/{owner}/{symbol}")
    @GET
    @Produces(MediaType.APPLICATION_JSON)
    @APIResponses(value = {
        @APIResponse(
            responseCode = "404",
            description = "The Mongo database cannot be found.",
            content = @Content(
                mediaType = "text/plain")),
        @APIResponse(
            responseCode = "200",
            description = "The ROI for the requested owner and symbol have been retrieved successfully.",
            content = @Content(
                mediaType = "application/json",
                schema = @Schema(implementation = Share.class)))})
    @Operation(summary = "Get the number of shares owned by specified owner for a specified stock symbol.")
    @WithSpan
    public Share getCurrentShares(
        @Parameter(description = "Owner name", required = true) @PathParam("owner") String ownerName,
        @Parameter(description = "Symbol name", required = false) @PathParam("symbol") String symbol) {

        Share shares = null;

        if (mConnector != null) {
            shares = mConnector.getSymbolShares(ownerName, symbol);
        }

        return shares;
    }

    @Path("/shares/{owner}")
    @GET
    @Produces(MediaType.APPLICATION_JSON)
    @APIResponses(value = {
        @APIResponse(
            responseCode = "404",
            description = "The Mongo database cannot be found.",
            content = @Content(
                mediaType = "text/plain")),
        @APIResponse(
            responseCode = "200",
            description = "The shares for the requested owner and symbol have been retrieved successfully.",
            content = @Content(
                mediaType = "application/json",
                schema = @Schema(implementation = Shares.class)))})
    @Operation(summary = "Get the number of shares of all owned stock by specified owner.")
    @WithSpan
    public Shares getPortfolioShares(@Parameter(description = "Owner name", required = true) @PathParam("owner") String ownerName) {

        Shares shares = null;

        if (mConnector != null) {
            shares = mConnector.getPortfolioShares(ownerName);
        }

        return shares;
    }

    @Path("/notional/{owner}")
    @GET
    @Produces(MediaType.APPLICATION_JSON)
    @APIResponses(value = {
        @APIResponse(
            responseCode = "404",
            description = "The Mongo database cannot be found. ",
            content = @Content(
                mediaType = "text/plain")),
        @APIResponse(
            responseCode = "200",
            description = "The notional for the requested owner and symbol have been retrieved successfully.",
            content = @Content(
                mediaType = "application/json",
                schema = @Schema(implementation = Double.class)))})
    @WithSpan
    public Double getNotional(
        @Parameter(description = "Owner name", required = true) @PathParam("owner") String ownerName) {

        Double notional = 0.0;

        if (mConnector != null) notional = mConnector.getTotalNotional(ownerName);

        return notional;
    }

    @Path("/returns/{owner}")
    @GET
    @Produces(MediaType.TEXT_PLAIN)
    @APIResponses(value = {
        @APIResponse(
            responseCode = "404",
            description = "The Mongo database cannot be found.",
            content = @Content(
                mediaType = "text/plain")),
        @APIResponse(
            responseCode = "200",
            description = "The ROI for the requested owner has been retrieved successfully.",
            content = @Content(
                mediaType = "application/json",
                schema = @Schema(implementation = Double.class)))})
    @Operation(summary = "Get the percentage return on portfolio for the specified owner, with passed in portfolio value.")
    @WithSpan
    public Double getReturns(
        @Parameter(description = "Owner name", required = true) @PathParam("owner") String ownerName,
        @Parameter(description = "Current portfolio value", required = true) @QueryParam("currentValue") Double portfolioValue) {

        String roi = "Unknown";

        if (mConnector != null) roi = mConnector.getROI(ownerName, portfolioValue);

        return Double.parseDouble(roi);
    }
}